Store automatic Answers JWT token to local storage.
To enable persistent automatic answers embed, we are storing the token
with expiry(a day from now) in the local storage. If the localStorage is
set the embed will show up until the local Storage expires. If the user
visits the article in help center via email with JWT token, then we will
reset the expiry to 24 hours from now.