Refactor slider navigation logic
Upon implementing the left tabbing support for the quick replies slider,
it was noticed that the existing logic is a bit all over the place
(e.g. need specific logic to handle directly focusing on the first item, need
different logic to handle directly focusing on the last item, different
logic to handle focusing on an item in the next/prev page). Due to its
complexity, it is harder to maintain and susceptible to break when the
logic is updated.

The logic is refactored with the idea of utilizing a single function (i.e.
goTo) to handle all navigation cases (i.e. via focusing on an item and
clicking the next/prev arrow button). Aside from that, the refactor also
improve the performance by doing the calculation once during the
component mounting.