[TALK-4008] Encode query params when connecting to talk embeddables service
We detected a problem with some customers where they weren't able to
display the talk integration in the web widget when they were using
nicknames with the "&" (ampersand) character.

Since we use the nickname as a query string param to connect to the
talk embeddables service, we need to ensure it's properly encoded in the
connection URL. This will fix any existing nickname with the "&" in it.