Add tooltip to structured message's button item
By design, structured message's button item can only have 1 line of text.
In case when the button text is longer than the maximum button width,
the text will be ellipsis-ed. But, once it is ellipsis-ed there will be
no way for end-user to know what's the full text is. Discussed with the
i18n engineer and concluded that a tooltip will be sufficient for this
case