Fix onChatStarted API callback fires twice bug
onChatStarted was firing twice when the page was being refreshed. This is becuase the Chat log was replaying the events on page load. The fix was to stop listening to the SDK events and instead listen to our own user events. Thus, it now listens to CHAT_MSG_REQUEST_SUCCESS instead.