Add browser browser test spec yml
After deciding to use the puppeteer framework as a scaffold to re-write a lot of our old browser tests and integrate them into this repo so they could be run in Travis before merging to master, we needed to figure out which of the old tests to port over. I was given this task and I decided to start out by writing a yml document of pseudocode test specs as I went. I reasoned that we needed high-level human-readable documentation of what the specifications should be, but also have a simple, easy way of viewing gaps in the coverage. A yml file provides this because it's easy to read and can be parsed, printed, counted and easily manipulated. It also provides a simple hierarchical structure that reflects the way the tests ought to be written out and reduces boilerplate in the documentation.