Set focus to search field whenever Help Center search field in rendered.
Removed comments

Re-implement set focus on search field when launcher is clicked.

Call the afterShowAnimate function after the refs is availabled in Frame.js
Create a timer ID for the setTimeOut function so that we can
kill the timer when the component unmounts.

Rename timer ID  and define it inside the constructor.

Set the reference to search field before typing on the keyboard
so that the search field component renders first before typing.

Remove `window`