Fix overflowing carousel in iOS12
Customer reported that the Web Widget is not rendered properly when
using iOS 12 and carousel is being rendered. There seems to be a iframe
bug in iOS 12. In iOS 12, if an iframe contain element that is wider
than iframe width, the iframe will be stretched until it can contain
children elements. A small reproduction can be seen at
https://jsbin.com/dejakaxete/edit?html,css,output. To fix this, the chat container max-width is set to 100vw to prevent the iframe element getting stretched.