Fix quick replies in history
In chat history, quick replies is rendered as fallback messages.
This is because of the wat quick replies is being rendered right now. It
will only be rendered properly for structured message that came from
SDK_CHAT_MSG action, but not others. This is because in chat-chats
reducer, in the event of SDK_CHAT_MSG, the chat object is manipulated
by replacing the msg with the one in structured_msg property. This
manipulation is not done for chat history that contains quick replies.
To fix it, moving this logic to the view layer that can be applied for
all scenarios.

Reference:
* Z1 ticket: https://support.zendesk.com/agent/tickets/4835107