Force update carousel on orientation change
This is to fix https://support.zendesk.com/agent/tickets/4836748 where
customer complains that the structured message carousel does not behave
properly in iOS device. It is caused by < iOS 13.2 devices does not
support ResizeObserver. Even though the carousel library that we use
(react-slick) do use resize-observer-polyfill, the polyfill does not
work in WebWidget due to the problem described in
https://zendesk.slack.com/archives/C0R1EJ3UP/p1574040302227800.

As a workaround, need to ensure SliderContainer component get updated on
orientation change.