Remove dog icon from channel choice
The last PR checked talkCapability rather than  talkCapability === CLICK_TO_CALL. This meant we would have shown the dog icon to customers, regardless of which talk option they add selected.

I've also changed the dog icon to a generic talk icon, since we've shown that it's possible for this icon to reach customers.