Change devServer host
This change is done to help us bring our dev environment more inline with prod. Specifically, when we tried to load the widget scripts into an iframe in dev we ran into Uncaught SyntaxError: The URL 'about:0.0.0.0:1337/sockjs-node' is invalid. Changing the host got us around this issue.

Since we don't have any strong need for the devserver to be publicly accessible this is considered an acceptable change.